refactor(cockpit): extract artifact writers
